Turns out if you're dealing with content that is 2GB or larger, you can fail to put the content into the cache because of this overflow. I couldn't find any reason why this cast was being made, so I'll just assume it was an oversight during development.

+++ Public/Src/Cache/ContentStore/Distributed/Blob/AzureBlobStorageContentSession.cs
@@ -624,7 +624,7 @@ namespace BuildXL.Cache.ContentStore.Distributed.Blobs
                         contentAlreadyExistsInCache = true;
                     }
 
-                    return new PutResult(contentHash, (int)contentSize, contentAlreadyExistsInCache);
+                    return new PutResult(contentHash, contentSize, contentAlreadyExistsInCache);
                 },
                 traceOperationStarted: false,
                 timeout: _configuration.StorageInteractionTimeout);
